requirements.txt
# This file tells Render which Python libraries to install.
# By not specifying a version, the latest one will be installed.
Flask
Pillow
google-generativeai
gunicorn